Initial work on simplifying the media library upload logic.
Based heavily off the work done by @alxy in #4314
